aplicacion_movil/backend/requirements.txt
fastapi
uvicorn
tensorflow
pillow
python-multipart
openai
